Always notify the client of OOB data
The reason for this change is heavily tied to the CTS as a user of
the client API. The CTS waits until it sees a PPS *and* and SPS NAL
before it starts sending new session parameters into the Vulkan API.

The "bug" is seen on the h264_resolution_change test, containing 30
frames. After the 15th one, a new SPS is correctly seen and processed
by Gst, it contains the resolution change information and is different
to the last SPS. Gst diligently forwards it on. Then comes a paired,
but technically redundant PPS. Gst notices it hasn't changed since the
last PPS and doesn't forward it on. However, that causes decoding to
continue before the CTS has gone ahead and updated the session
parameters with the new info in the SPS (since it's still waiting for
a redundant PPS)

The pushing of OOB data only when it's really changed makes sense to
me, and perhaps it's even spec'd like that. I didn't dig deeper to get
definitive answer to what the behaviour should be here, but
unfortunately it's the behaviour of the Nvidia parser, and we should
probably follow that if we have the choice. Changing the CTS handling
would require getting that proprietary parser to also detect redundant
parameter objects, and then to adjust the CTS to update more often.